score: allow to disable scoring of objects in specific namespaces
This features adds the ability to disable tests on the specified
namespaces.

Specific usage example:

```
kube-score score --ignore-namespace=logging,minio -
```

This usecase is specifically for ignoring the checks on namespace like
istio-system which isn't managed by a specific team.

```
RELNOTE: score: allow to disable scoring of objects in specific namespaces
```