score/disruptionbudget: output comment when namespace doesn't match
I was troubleshooting a strange kube-score failure regarding
PodDisruptionBudget.

kube-score was complaining that a deployment didn't have a PDB even
though it very much did.

Going back to the code I realized the test was exiting early due to the
namespace check:

https://github.com/zegl/kube-score/blob/0b3f154ca3f06a13323431a7d2199a74a1869fbc/score/disruptionbudget/disruptionbudget.go#L23-L25

Obviously, this was a bad manifest on my end (explicit is better than
implicit and the tests passed as soon as I added `namespace: {{
.Release.Namespace }}` to my helm charts, but perhaps with a bit more
output we could help other folks connect the dots faster.